+++ src/utils/text.ts
@@ -17,12 +17,20 @@ export function warningText(string: string): string {
   return chalk.yellow(string)
 }
 
+export function errorText(string: string): string {
+  return chalk.red(string)
+}
+
 export function deletePreviousLine(): void {
   process.stdout.write('\r' + goUpOneRow() + deleteWholeRow())
 }
 
 export function createKeyValue(key: string, value: string | number | boolean, padLength?: number): string {
   return `${chalk.green.bold(key + ':').padEnd(padLength ? padLength + 1 : 0)} ${String(value)}`
+}
+
+export function ellipsis(value: string, startIndex: number, endIndex?: number | null) {
+  return `${value.slice(0, startIndex)}...${endIndex ? value.slice(endIndex) : ''}`
 }
 
 export function printDivided<T>(